+++ tradedangerous/transfers.py
@@ -103,6 +103,7 @@ def download(
             backup=False,
             shebang=None,
             chunkSize=4096,
+            timeout=90,
         ):
     """
     Fetch data from a URL and save the output
@@ -126,7 +127,7 @@ def download(
     
     requests = import_requests()  # pylint: disable=redefined-outer-name
     tdenv.NOTE("Requesting {}".format(url))
-    req = requests.get(url, headers=headers or None, stream=True, timeout=300)
+    req = requests.get(url, headers=headers or None, stream=True, timeout=timeout)
     req.raise_for_status()
     
     encoding = req.headers.get('content-encoding', 'uncompress')
@@ -142,6 +143,14 @@ def download(
             raise TradeException(
                 "Remote server gave an empty response. Please try again later."
             )
+
+    # if the file is being compressed by the server, the headers tell us the
+    # length of the compressed data, but in our loop below we will be receiving
+    # the uncompressed data, which should be larger, which will cause our
+    # download indicators to sit at 100% for a really long time if the file is
+    # heavily compressed and large (e.g spansh 1.5gb compressed vs 9GB uncompressed)
+    if encoding == "gzip" and length:
+        length *= 4
     
     if tdenv.detail > 1:
         if length:
@@ -222,7 +231,7 @@ def download(
     req.close()
     return req.headers
 
-def get_json_data(url):
+def get_json_data(url, *, timeout: int = 90):
     """
     Fetch JSON data from a URL and return the resulting dictionary.
     
@@ -230,7 +239,7 @@ def get_json_data(url):
     """
     
     requests = import_requests()  # pylint: disable=redefined-outer-name
-    req = requests.get(url, stream=True, timeout=300)
+    req = requests.get(url, stream=True, timeout=timeout)
     
     totalLength = req.headers.get('content-length')
     if totalLength is None:
@@ -266,12 +275,12 @@ class CSVStream:
             print("{} = {}".format(cols[0], vals[0]))
     """
     
-    def __init__(self, url, tdenv=None):
+    def __init__(self, url, tdenv=None, *, timeout: int = 90):
         self.url = url
         self.tdenv = tdenv
         if not url.startswith("file:///"):
             requests = import_requests()  # pylint: disable=redefined-outer-name
-            self.req = requests.get(self.url, stream=True, timeout=300)
+            self.req = requests.get(self.url, stream=True, timeout=timeout)
             self.lines = self.req.iter_lines()
         else:
             self.lines = open(url[8:], "rUb")